make activityrow its own thing
