--- src/components/activity/activity.json
@@ -1,66 +1,0 @@
-[
-    {
-        "obj_id": 82475328, 
-        "datetime_created": "2015-10-20T15:13:36", 
-        "actor": {
-            "username": "ceebee", 
-            "pk": 2755634, 
-            "thumbnail_url": "//cdn.scratch.mit.edu/static/site/users/avatars/275/5634.png", 
-            "admin": true
-        }, 
-        "pk": 186757838, 
-        "message": "\nfavorited\n   <a href=\"/projects/82475328/\">miner man</a>", 
-        "extra_data": {
-            "project_title": "miner man"
-        }, 
-        "type": 3
-    }, 
-    {
-        "obj_id": 82475328, 
-        "datetime_created": "2015-10-20T15:13:36", 
-        "actor": {
-            "username": "ceebee", 
-            "pk": 2755634, 
-            "thumbnail_url": "//cdn.scratch.mit.edu/static/site/users/avatars/275/5634.png", 
-            "admin": true
-        }, 
-        "pk": 186757836, 
-        "message": "\nloved\n    <a href=\"/projects/82475328/\">miner man</a>", 
-        "extra_data": {
-            "project_title": "miner man"
-        }, 
-        "type": 2
-    }, 
-    {
-        "obj_id": 82475328, 
-        "datetime_created": "2015-10-20T15:12:39", 
-        "actor": {
-            "username": "speakvisually", 
-            "pk": 3484484, 
-            "thumbnail_url": "//cdn.scratch.mit.edu/static/site/users/avatars/348/4484.png", 
-            "admin": true
-        }, 
-        "pk": 186757510, 
-        "message": "\nfavorited\n   <a href=\"/projects/82475328/\">miner man</a>", 
-        "extra_data": {
-            "project_title": "miner man"
-        }, 
-        "type": 3
-    }, 
-    {
-        "obj_id": 82475328, 
-        "datetime_created": "2015-10-20T15:12:37", 
-        "actor": {
-            "username": "speakvisually", 
-            "pk": 3484484, 
-            "thumbnail_url": "//cdn.scratch.mit.edu/static/site/users/avatars/348/4484.png", 
-            "admin": true
-        }, 
-        "pk": 186757500, 
-        "message": "\nloved\n    <a href=\"/projects/82475328/\">miner man</a>", 
-        "extra_data": {
-            "project_title": "miner man"
-        }, 
-        "type": 2
-    }
-]
