+++ src/views/splash/splash.scss
@@ -66,6 +66,14 @@
         }
     }
 
+    .activity {
+        ul {
+            display: block;
+            margin: 0;
+            padding: 0;
+        }
+    }
+
 
     .box {
         margin-bottom: 20px;
